Passing rust a python generator instead of a list
This means we don't have to load the entire dataset into memory. instead
we can iteratively count duplicate blocks and then only train on unique
blocks thus saving memory. (This was already partly implemented but we
were still counting duplicate blocks from the full dataset loaded into
memory which was pointless as we never used the full dataset after that).