Reducing to rank 0 before merging pairs
The main bottleneck is counting blocks. If we do this across multiple
processes we can get a big speed up, but then merging is slowed down a
lot because it is very communication heavy.

To get the best of both worlds we count across multiple processes and
then reduce to rank 0 for merging. This relies on being able to expand
the resources available to a single process (rank 0) mid way through
execution while maintaining all the information gathered from other
ranks.